Moving TransformationPipeline to model
